requirements.txt
flask==0.11.1
pysensu==0.5
PyYAML==3.11